Clean up TaskPropertiesTests  - Shorten Task.ID uniqueness test.
Create options as combinatorial member data.

- Previously did not use option RunContinuationsAsynchronously

Move DeliberateTestException to Common folder

Pull flag out to common file
