--- src/System.Linq.Parallel/tests/Helpers/DeliberateTestException.cs
@@ -1,9 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace System.Linq.Parallel.Tests
-{
-    internal class DeliberateTestException : Exception
-    {
-    }
-}
